# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação - API SuricatoDev (Projeto Excursionistas)",
        "description": "Esta API fornece endpoints para a gestão de caravanas, passageiros e organizadores, realizar reservas, denúncias e avaliações.\n *      🚀 **Principais funcionalidades:**\n *      - Gerenciar caravanas\n *      - Gerenciar Passageiros\n *      - Gerenciar organizadores\n *      - Realização de reservas\n *      - Denúncias e avaliações de caravanas, passageiros e organizadores\n *      🔒 **Segurança:**\n *      - Autenticação segura via token Bearer\n *      - Utiliza Laravel Sanctum para autenticação\n *      - Requer token Bearer para acesso a endpoints protegidos",
        "contact": {
            "name": "Suporte Suricato Dev",
            "email": "filipe.lamego@fatec.sp.gov.br"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.3.0"
    },
    "paths": {
...
        "/api/denuncias": {
            "post": {
                "tags": [
                    "Denúncias"
                ],
                "summary": "Registrar uma nova denúncia",
                "description": "Registra uma denúncia com base no tipo de usuário (passageiro ou organizador).",
                "operationId": "6f27675ace1640fa0086c81a59ee5663",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "denunciante_id",
                                    "topico",
                                    "descricao"
                                ],
                                "properties": {
                                    "denunciante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "topico": {
                                        "type": "string",
                                        "enum": [
                                            "Pagamento",
                                            "Cancelamento",
                                            "Inconsistência",
                                            "Segurança",
                                            "Outro"
                                        ],
                                        "example": "Pagamento"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "O pagamento não foi processado corretamente."
                                    },
                                    "caravana_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "organizador_id": {
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    },
                                    "passageiro_id": {
                                        "type": "integer",
                                        "example": 4,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Denúncia registrada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Denúncia registrada com sucesso!"
                                        },
                                        "denuncia": {
                                            "type": "object",
                                            "example": {
                                                "id": 10,
                                                "denunciante_id": 1,
                                                "topico": "Pagamento",
                                                "descricao": "O pagamento não foi processado corretamente.",
                                                "status": "Pendente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de validação."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "topico": [
                                                    "O campo tópico é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao registrar denúncia.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao registrar denúncia."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/denuncias/{id}": {
            "put": {
                "tags": [
                    "Denúncias"
                ],
                "summary": "Editar uma denúncia",
                "description": "Permite ao usuário editar sua própria denúncia.",
                "operationId": "a6aeb21ba9ca8692e4fd7a2397309e17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da denúncia a ser editada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "topico": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Inconsistência"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "A caravana foi alterada sem aviso prévio."
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Pendente",
                                            "Em andamento",
                                            "Concluído"
                                        ],
                                        "example": "Em andamento"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Denúncia atualizada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Denúncia atualizada com sucesso!"
                                        },
                                        "denuncia": {
                                            "type": "object",
                                            "example": {
                                                "id": 10,
                                                "denunciante_id": 1,
                                                "topico": "Inconsistência",
                                                "descricao": "A caravana foi alterada sem aviso prévio.",
                                                "status": "Em andamento"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuário não tem permissão para editar.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Você não tem permissão para editar esta denúncia."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar denúncia.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao atualizar denúncia."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Denúncias"
                ],
                "summary": "Excluir uma denúncia",
                "description": "Permite ao usuário excluir sua própria denúncia.",
                "operationId": "d9a54ca4d36b74cfc64aa969f3cc3c0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da denúncia a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Denúncia excluída com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Denúncia excluída com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuário não tem permissão para excluir.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Você não tem permissão para excluir esta denúncia."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao excluir denúncia.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao excluir denúncia."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Autenticação via Bearer Token usando Laravel Sanctum"
            }
        }
    },
...
}